+++ CHANGELOG.md
@@ -10,11 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Config accepts "clazzfish.dump.uri" as additional system property
+- import and export methods in ClazzStatistic accepts URI as parameter
+- CsvXPorterProvider can now registered for import and export from and to an URI
 
 ### Changed
 - array types are filtered out
   ([issue #23](https://github.com/oboehm/ClazzFish/issues/23))
 - ClazzStatistic writes the statistic not direct but to a tempory file to be more robust for parallel exports
+
+### Breaking Changes
+- exportCSV-Methods in ClazzStatistic returns now an URI (before: File)
+
 
 ## [2.4.2] (10-Jan-2025)